# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "AccountHolderResponse": {
        "properties": {
          "holderType": {
            "type": "string"
          },
          "memberId": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "holderType",
          "memberId"
        ],
        "type": "object"
      },
      "AccountResponse": {
        "properties": {
          "accountKind": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "hasFullNumber": {
            "type": "boolean"
          },
          "holders": {
            "items": {
              "$ref": "#/components/schemas/AccountHolderResponse"
            },
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "ifsc": {
            "type": "string"
          },
          "institutionId": {
            "format": "uuid",
            "type": "string"
          },
          "institutionName": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "linkedInvestmentCount": {
            "format": "int32",
            "type": "integer"
          },
          "notes": {
            "type": "string"
          },
          "numberMasked": {
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          },
          "visibility": {
            "type": "string"
          },
          "visibleToMemberIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "accountKind",
          "createdAt",
          "hasFullNumber",
          "holders",
          "id",
          "label",
          "linkedInvestmentCount",
          "version",
          "visibility",
          "visibleToMemberIds"
        ],
        "type": "object"
      },
...
      "DocumentResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "docType": {
            "type": "string"
          },
          "expired": {
            "type": "boolean"
          },
          "expiresOn": {
            "format": "date",
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/DocumentLinkResponse"
            },
            "type": "array"
          },
          "mimeType": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "sizeBytes": {
            "format": "int64",
            "type": "integer"
          },
          "visibility": {
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "docType",
          "expired",
          "fileName",
          "id",
          "links",
          "mimeType",
          "sizeBytes",
          "visibility"
        ],
        "type": "object"
      },
      "DownloadTicket": {
        "properties": {
          "expiresInSeconds": {
            "format": "int64",
            "type": "integer"
          },
          "fileName": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "expiresInSeconds",
          "fileName",
          "token"
        ],
        "type": "object"
      },
...
      "HolderInput": {
        "properties": {
          "holderType": {
            "type": "string"
          },
          "memberId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "holderType",
          "memberId"
        ],
        "type": "object"
      },
...
      "InvestmentResponse": {
        "properties": {
          "accountId": {
            "format": "uuid",
            "type": "string"
          },
          "accountLabel": {
            "type": "string"
          },
          "attributes": {
            "additionalProperties": {},
            "type": "object"
          },
          "categoryCode": {
            "type": "string"
          },
          "categoryLabel": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "encumbrance": {
            "type": "number"
          },
          "encumbranceFormatted": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "institutionId": {
            "format": "uuid",
            "type": "string"
          },
          "institutionName": {
            "type": "string"
          },
          "investedAmount": {
            "type": "number"
          },
          "isInContinuity": {
            "type": "boolean"
          },
          "lastVerifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "maturityDate": {
            "format": "date",
            "type": "string"
          },
          "netEquity": {
            "type": "number"
          },
          "nominees": {
            "items": {
              "$ref": "#/components/schemas/NomineeResponse"
            },
            "type": "array"
          },
          "notes": {
            "type": "string"
          },
          "owners": {
            "items": {
              "$ref": "#/components/schemas/OwnerResponse"
            },
            "type": "array"
          },
          "quantity": {
            "type": "number"
          },
          "startDate": {
            "format": "date",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "storageLocation": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "typeCode": {
            "type": "string"
          },
          "typeIcon": {
            "type": "string"
          },
          "typeId": {
            "format": "uuid",
            "type": "string"
          },
          "typeLabel": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "valueBasis": {
            "type": "string"
          },
          "valueFormatted": {
            "type": "string"
          },
          "valuedOn": {
            "format": "date",
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          },
          "visibility": {
            "type": "string"
          },
          "visibleToMemberIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "attributes",
          "categoryCode",
          "categoryLabel",
          "color",
          "createdAt",
          "currency",
          "id",
          "isInContinuity",
          "nominees",
          "owners",
          "status",
          "title",
          "typeCode",
          "typeId",
          "typeLabel",
          "valueBasis",
          "version",
          "visibility",
          "visibleToMemberIds"
        ],
        "type": "object"
      },
...
      "MissingProof": {
        "properties": {
          "investmentId": {
            "format": "uuid",
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "investmentId",
          "title"
        ],
        "type": "object"
      },
...
      "UpdateAccount": {
        "properties": {
          "holders": {
            "items": {
              "$ref": "#/components/schemas/HolderInput"
            },
            "type": "array"
          },
          "ifsc": {
            "type": "string"
          },
          "institutionId": {
            "format": "uuid",
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "storeFullNumber": {
            "type": "boolean"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Access token from POST /api/v1/auth/otp/verify.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "padmini.koduri@bhrigu.tech",
      "name": "Almira"
    },
    "description": "Every rupee a family has invested or owes, in one private place.\n\n**Authentication.** Phone OTP issues a short-lived access token and a\nrotating refresh token. Send `Authorization: Bearer <accessToken>`.\nRefresh tokens are single-use: presenting one twice is treated as theft\nand ends the session.\n\n**Privacy.** Every read passes a per-record visibility check enforced by\nPostgreSQL row-level security. A record you may not see returns 404, not\n403 — a 403 would confirm it exists. Totals are computed through the same\nfilter, so a private record contributes nothing, not even its amount, to\nanother member's figures.\n\n**Money** is decimal, never floating point. Amounts also arrive formatted\n(`₹1,76,875`) so every surface renders them identically.\n\n**Concurrency.** Writes carry the record's `version`; a stale write is\nrejected with 409 and the current version. Creates accept a client-supplied\n`id`, so an offline capture keeps its identity and a retry is idempotent.\n\n**Stability.** v1 is frozen and additive-only. Breaking changes go to v2.",
    "title": "Almira API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/download": {
      "get": {
        "operationId": "download",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "document-download-controller"
        ]
      }
    },
...
    "/api/v1/households/{householdId}/accounts/{id}": {
      "delete": {
        "operationId": "archive_2",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "account-controller"
        ]
      },
      "get": {
        "operationId": "get_3",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "account-controller"
        ]
      },
      "patch": {
        "operationId": "update_3",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccount"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "account-controller"
        ]
      }
    },
...
    "/api/v1/households/{householdId}/documents": {
      "get": {
        "operationId": "list_4",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "entityType",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "entityId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "document-controller"
        ]
      },
      "post": {
        "operationId": "upload",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "docType",
            "required": false,
            "schema": {
              "default": "other",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expiresOn",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "visibility",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "notes",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "entityType",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "entityId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "document-controller"
        ]
      }
    },
    "/api/v1/households/{householdId}/documents/missing-proof": {
      "get": {
        "operationId": "missingProof",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MissingProof"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "document-controller"
        ]
      }
    },
    "/api/v1/households/{householdId}/documents/{id}": {
      "delete": {
        "operationId": "archive_3",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "document-controller"
        ]
      },
      "get": {
        "operationId": "get_4",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "document-controller"
        ]
      }
    },
    "/api/v1/households/{householdId}/documents/{id}/access": {
      "post": {
        "operationId": "requestAccess",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadTicket"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "document-controller"
        ]
      }
    },
    "/api/v1/households/{householdId}/documents/{id}/links": {
      "delete": {
        "operationId": "unlink",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "entityType",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "entityId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "document-controller"
        ]
      },
      "post": {
        "operationId": "link",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "entityType",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "entityId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "document-controller"
        ]
      }
    },
...
    "/api/v1/households/{householdId}/investments/{id}": {
      "delete": {
        "operationId": "archive_1",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "investment-controller"
        ]
      },
      "get": {
        "operationId": "get_2",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "investment-controller"
        ]
      },
      "patch": {
        "operationId": "update_2",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvestment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "investment-controller"
        ]
      }
    },
...
    "/api/v1/households/{householdId}/liabilities/{id}": {
      "delete": {
        "operationId": "archive",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "liability-controller"
        ]
      },
      "get": {
        "operationId": "get_1",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LiabilityResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "liability-controller"
        ]
      },
      "patch": {
        "operationId": "update_1",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLiability"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LiabilityResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "liability-controller"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ]
}